requirements.txt
pynput
Pillow
transformers
schedule
tf-keras
pystray
pillow
tkinter